Implement PDF export for Budget reports

+++ src/main/java/com/example/intelliwealth/treasury/budget/infrastructure/persistence/BudgetRepository.java
@@ -1,7 +1,10 @@
 package com.example.intelliwealth.treasury.budget.infrastructure.persistence;
 
+import com.example.intelliwealth.treasury.budget.application.dto.BudgetSummaryDTO;
 import com.example.intelliwealth.treasury.budget.domain.model.Budget;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
@@ -9,7 +12,9 @@ import java.util.Optional;
 import java.util.UUID;
 
 @Repository
-public interface BudgetRepository extends JpaRepository<Budget, Integer> {
+public interface BudgetRepository extends JpaRepository<Budget, Long> {
+
     List<Budget> findAllByUserId(UUID userId);
-    Optional<Budget> findByIdAndUserId(int id, UUID userId);
+
+    Optional<Budget> findByIdAndUserId(Long id, UUID userId);
 }
